+++ NEWS.md
@@ -4,6 +4,8 @@
 * New pipes `%P.%`, `%<P.%`, `%V.%`,  and `%<V.%`, are introduced to `print()` or
 `View()` the result of every step.
 * We can use `if` and `else` to have conditional steps
+* `%lapply..%` and `%sapply..%` are counterparts to `%lapply.%` and `%sapply.%`
+* `%F.%` and `%F..%` reproduce magrittr's functional sequence feature. 
 
 # nakedpipe 0.0.0.9000